backend/requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
defusedxml==0.5.0
Django==2.2
django-allauth==0.39.1
django-cors-headers==2.5.2
django-rest-auth==0.9.5
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
psycopg2-binary==2.8.2
pycparser==2.19
PyJWT==1.7.1
python-binary-memcached==0.28.0
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3
